README.md
Step1. Merges the training and the test sets to create one data set.
Step2. Extracts only the measurements on the mean and standard deviation for each measurement.
Step 3. Use descriptive activity names for naming the data set. Then label the data set with activity names.
Step 4. Create a tidy data set with the average of each variable for each activity and each subject.